requirements.txt
fastapi
uvicorn
jinja2
google-genai
google-adk
